CI: Allow for generated code in PRs
Generated code must now be included in PRs. This is helpful during
review, as it provides the reviewer with the ability to see the result
of a change.

This is also necessary after https://github.com/hemtjanst/zyaml/commit/78cfbb9dfc235b79a615a83e412052e7edfa6a12
as it'll now raise PRs instead, which will include the generated code.